pre-deploy snapshot

+++ skills/nocobase-dsl-reconciler/workspaces/erp/pages/erp/dashboard/charts/inventory_by_category.sql
@@ -1,6 +1,6 @@
 SELECT
   COALESCE(category, 'Uncategorized') AS label,
-  ROUND(COALESCE(SUM(current_stock * standard_cost), 0), 2) AS value
+  ROUND(COALESCE(SUM(current_stock * standard_cost), 0)::numeric, 2) AS value
 FROM nb_erp_products
 GROUP BY COALESCE(category, 'Uncategorized')
 ORDER BY value DESC, label ASC;
