+++ skills/nocobase-dsl-reconciler/workspaces/erp/pages/erp/sales_analytics/charts/revenue_by_segment.sql
@@ -1,6 +1,6 @@
 SELECT
   COALESCE(c.segment, 'unclassified') AS label,
-  ROUND(COALESCE(SUM(o.total_amount), 0), 2) AS value
+  ROUND(COALESCE(SUM(o.total_amount), 0)::numeric, 2) AS value
 FROM nb_erp_customer_orders o
 LEFT JOIN nb_erp_customers c ON o."customerId" = c.id
 WHERE o.status <> 'cancelled'
