+++ skills/nocobase-dsl-reconciler/workspaces/erp/pages/erp/sales_analytics/charts/revenue_trend.sql
@@ -1,6 +1,6 @@
 SELECT
   TO_CHAR(DATE_TRUNC('month', order_date::timestamp), 'YYYY-MM') AS label,
-  ROUND(COALESCE(SUM(total_amount), 0), 2) AS value
+  ROUND(COALESCE(SUM(total_amount), 0)::numeric, 2) AS value
 FROM nb_erp_customer_orders
 WHERE status <> 'cancelled'
 GROUP BY DATE_TRUNC('month', order_date::timestamp)
